Implemented PreparedStatement for sqlite
Expanded the PreparedStatement baseclass so that it is responsible for
managing bound variables. This needs more testing though, making sure
there are no memory leaks.

Implemented the SqlitePreparedStatement subclass and wired it into the
SqliteDatabaseHandler implementation.

Wrote a few basic tests for PreparedStatement. These needs to be expanded
first.

Created a sqlite fixture database. Later on something along similar
lines will have to be created for mysql. For sqlite we can simply
generate a database and copy/delete it for each test. For mysql however
there will be a need to create/drop the entire testing database. This
will be much more time-consuming but is a required necessity so that
tests have an ensured database-state at the beginning ot the test.